Cleanup dependencies & unify http client impl
