+++ Cargo.toml
@@ -16,3 +16,5 @@ reqwest = { version = "0.12.9", default-features = false, features = ["stream", 
 test-log = { version = "0.2.14", default-features = false, features = ["trace"] }
 tokio = { version = "1.43.1", features = ["rt-multi-thread"] }
 shadow-rs = "0.32.0"
+thiserror = "1.0.63"
+http = "1"
