+++ tsumugu-parser/Cargo.toml
@@ -14,14 +14,14 @@ tracing = { workspace = true }
 clap = { workspace = true }
 chrono = { workspace = true }
 url = { workspace = true }
+thiserror = { workspace = true }
+http = { workspace = true }
 
 flate2 = "1.0.28"
 apt-parser = "1.0.0"
 percent-encoding = "2.3.1"
 regex = "1.9.1"
-thiserror = "1.0.63"
 scraper = "0.23.1"
-http = "1"
 
 [dev-dependencies]
 test-log = { workspace= true }
